Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
nltk
numpy